background_remover/requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
numpy<2
